+++ packages/maplibre/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackwright/maplibre
 
+## 2.0.3-alpha.1
+
+### Patch Changes
+
+- @stackwright/core@0.8.3-alpha.1
+
+## 2.0.3-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [5cfa88e]
+  - @stackwright/core@0.8.3-alpha.0
+
 ## 2.0.2
 
 ### Patch Changes